requirements.txt
anthropic
python-telegram-bot
google-auth-oauthlib
google-api-python-client
gspread
slack-sdk
notion-client
